# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Eternal Kingdoms API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /worlds/{worldId}/resource-nodes:
    get:
      operationId: getResourceNodes
      tags: [worlds]
      summary: Get active resource nodes in a world
      parameters:
        - name: worldId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Active resource nodes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceNodesResponse"
  /worlds/{worldId}/monster-nodes:
    get:
      operationId: getMonsterNodes
      tags: [worlds]
      summary: Get active monster spawn nodes in a world
      parameters:
        - name: worldId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Active monster nodes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonsterNodesResponse"

  # ── Phase 4: Heroes ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ResourceNodeDto:
      type: object
      required: [id, worldId, spawnType, spawnSubtype, level, tileX, tileY, posX, posY, biome, status, spawnedAt, expiresAt]
      properties:
        id:
          type: integer
        worldId:
          type: integer
        spawnType:
          type: string
        spawnSubtype:
          type: string
        level:
          type: integer
        tileX:
          type: integer
        tileY:
          type: integer
        posX:
          type: integer
        posY:
          type: integer
        biome:
          type: string
        status:
          type: string
        spawnedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    ResourceNodesResponse:
      type: object
      required: [worldId, nodes]
      properties:
        worldId:
          type: integer
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/ResourceNodeDto"
    MonsterNodeDto:
      type: object
      required: [id, worldId, spawnType, spawnSubtype, level, tileX, tileY, posX, posY, biome, status, spawnedAt, expiresAt]
      properties:
        id:
          type: integer
        worldId:
          type: integer
        spawnType:
          type: string
        spawnSubtype:
          type: string
        level:
          type: integer
        tileX:
          type: integer
        tileY:
          type: integer
        posX:
          type: integer
        posY:
          type: integer
        biome:
          type: string
        status:
          type: string
        spawnedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    MonsterNodesResponse:
      type: object
      required: [worldId, monsters]
      properties:
        worldId:
          type: integer
        monsters:
          type: array
          items:
            $ref: "#/components/schemas/MonsterNodeDto"

    # ── Phase 4 Schemas ────────────────────────────────────────────────────────
...