# lib/api-zod/src/generated/types/resourceNodeDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Eternal Kingdoms API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceNodeDto {
  id: number;
  worldId: number;
  spawnType: string;
  spawnSubtype: string;
  level: number;
  tileX: number;
  tileY: number;
  posX: number;
  posY: number;
  biome: string;
  status: string;
  spawnedAt: Date;
  expiresAt: Date;
}