# lib/api-zod/src/generated/types/resourceNodesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Eternal Kingdoms API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceNodeDto } from './resourceNodeDto';

export interface ResourceNodesResponse {
  worldId: number;
  nodes: ResourceNodeDto[];
}